pyproject.toml
[tool.poetry]
name = "whatsapp-le-parser"
version = "0.1.0"
description = "Toolkit for parsing Whatsapp law enforcement data requests"
authors = ["Your Name <you@example.com>"]
license = "WTFPLv2"
readme = "README"
packages = [{include = "parser"}]

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.12.2"
phonenumbers = "^8.13.20"
python-dateutil = "^2.8.2"


[tool.poetry.scripts]
whatsapp-le-parser = "parser:cli"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
